CDR-2305 extract TemplateCacheHelper
